django-practice/src/polls/api/serializers.py
from rest_framework import serializers
from polls.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ["title", "description", "is_active"]
